+++ runtime/default.nix
@@ -19,7 +19,7 @@
 }:
 mkDerivation {
   pname = "hoppy-runtime";
-  version = "0.6.0";
+  version = "0.7.0";
   src = ./.;
   libraryHaskellDepends = [ base ];
   homepage = "http://khumba.net/projects/hoppy";