+++ std/default.nix
@@ -20,7 +20,7 @@
 }:
 mkDerivation {
   pname = "hoppy-std";
-  version = "0.6.0";
+  version = "0.7.0";
   src = ./.;
   libraryHaskellDepends = [
     base filepath haskell-src hoppy-generator