+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## Unreleased
 
-- [API change] There is a new `EnumScoped` which is now expected as an argument
-  to `makeAutoEnum` rather than a boolean to indicate the scopedness of an enum.
+- [API change] There is a new `Scoped` which is now expected as an argument to
+  `makeAutoEnum` rather than a boolean to indicate the scopedness of an enum.
 
 - Fix for autodetection of scoped enum values (issue #43).  For now,
   autodetection for scoped enums only works with C++11 and up.